Modifed plugin to work more like delay and added test

+++ flexget/plugins/filter/pending_approval.py
@@ -4,6 +4,7 @@ from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 import logging
 
 from flexget.manager import Session
+from flexget.utils.database import entry_synonym
 from sqlalchemy import Column, String, Unicode, Boolean, Integer
 from flexget import db_schema, plugin
 from flexget.event import event
@@ -20,12 +21,15 @@ class PendingEntry(Base):
     title = Column(Unicode)
     url = Column(String)
     approved = Column(Boolean)
+    _json = Column('json', Unicode)
+    entry = entry_synonym('_json')
 
-    def __init__(self, task_name, title, url):
+    def __init__(self, task_name, entry):
         self.task_name = task_name
-        self.title = title
-        self.url = url
+        self.title = entry['title']
+        self.url = entry['url']
         self.approved = False
+        self.entry = entry
 
     def __repr__(self):
         return '<PendingEntry(task_name={},title={},url={},approved={})>' \
@@ -44,31 +48,44 @@ class PendingApproval(object):
             .first()
         return db_entry
 
-    @plugin.priority(0)
-    def on_task_filter(self, task, config):
+    @plugin.priority(-1)
+    def on_task_input(self, task, config):
         if not config:
             return
 
-        for entry in task.entries:
-            with Session() as session:
-                db_entry = self._item_query(entry, task, session)
-                if not db_entry:
-                    db_entry = PendingEntry(task_name=task.name, title=entry['title'], url=entry['url'])
-                    log.debug('creating new pending entry %s', db_entry)
-                session.merge(db_entry)
+        approved_entries = []
+        with Session() as session:
+            # Let details plugin know that it is ok if this task doesn't produce any entries
+            task.no_entries_ok = True
 
-                if db_entry.approved is True:
-                    entry.accept('approval was set to True')
+            for entry in task.entries:
+                # Cache all new task entries
+                if not self._item_query(entry, task, session):
+                    log.verbose('creating new pending entry %s', entry)
+                    session.add(PendingEntry(task_name=task.name, entry=entry))
+
+            # Clear the current entries from the task now that they are stored
+            task.all_entries[:] = []
+
+            # Pass all entries marked as approved
+            for approved_entry in session.query(PendingEntry).filter(PendingEntry.approved == True).all():
+                e = approved_entry.entry
+                e['approved'] = True
+                approved_entries.append(e)
+
+        return approved_entries
 
     def on_task_learn(self, task, config):
         if not config:
             return
-        for entry in task.accepted:
-            with Session() as session:
-                db_entry = self._item_query(entry, task, session)
-                if db_entry:
-                    log.debug('deleting approved entry %s', db_entry)
-                    session.delete(db_entry)
+        with Session() as session:
+            # Delete all entries that have passed the pending phase
+            for entry in task.entries:
+                if entry.get('approved', False) is True:
+                    db_entry = self._item_query(entry, task, session)
+                    if db_entry and db_entry.approved:
+                        log.debug('deleting approved entry %s', db_entry)
+                        session.delete(db_entry)
 
 
 @event('plugin.register')
